requirements.txt
# Web Framework & Server
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-login==0.6.3
flask-wtf==1.1.1
flask-cors==4.0.0
flask-mail==0.9.1
werkzeug==2.3.7

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# AI/ML & Data Science
numpy==1.24.4
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1
joblib==1.3.2

# Ollama & AI Integration
ollama==0.1.7
openai==0.28.1
transformers==4.35.2
torch==2.1.0
sentence-transformers==2.2.2

# Document Processing & PDF
fpdf2==2.7.5
python-docx==1.1.0
reportlab==4.0.4
pillow==10.0.1
opencv-python==4.8.1.78
pytesseract==0.3.10

# Data Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0

# File Processing & Utilities
python-multipart==0.0.6
python-magic==0.4.27
pyyaml==6.0.1
requests==2.31.0
urllib3==1.26.18

# Security & Cryptography
cryptography==41.0.7
bcrypt==4.0.1
pyjwt==2.8.0

# Excel/CSV Processing
openpyxl==3.1.2
xlrd==2.0.1

# Testing & Development
pytest==7.4.2
pytest-cov==4.1.0
black==23.9.1
flake8==6.1.0
autopep8==2.0.4

# Environment & Configuration
python-dotenv==1.0.0
configparser==5.3.0

# Async & Performance
aiohttp==3.8.6
asyncio==3.4.3


# Date/Time
python-dateutil==2.8.2
pytz==2023.3

# CLI & Logging
click==8.1.7
colorama==0.4.6
loguru==0.7.2

# Additional Utilities
tqdm==4.66.1
chardet==5.2.0
beautifulsoup4==4.12.2
lxml==4.9.3